Fixed IDKEY in Secureform
I still had problems whith Security warnings on pages that used alot of
IDKEYS. For example if i tried to move a field in the basic form module
i had one  Warning on in about 15 atempts.

Another Problem: unused old IDKEYS where never deleted so session got
spammed by idkeys.

So i made a more unique ID, recursive checking for overlapping ids,
added a tiemout for IDKEYS (same as for FTAN) and added a page parameter
to instantly delete all keys from a pagecall if one is used.

Last thing i added is an additional parameter that allows the use of
IDKEYS whith ajax scripts. If this setting is used, the key is not
removed and can be used multiple times. (this is a small security risk
but nothing compared to the 1 in 16 chance to guess the key that we had
before. )

This is a patch from my old fork only tested on my local installation.