Changed language file loading mechanism
Now we load the default file (EN) and then we load the
local(userdefined) file. That way even an incomplete Language file at
least displays the english text instead of an error message.

Loading of compatibility file now moved to initialize instead of
doing it in each separated language file.