Adding a maintainance mode
The Maintainance mode is activated by setting a constant in the config.php.

define('WB_MAINTAINANCE_MODE', true);

You can deactivate it by either deleting the line or by setting the mode
to false.

The default template for maintainance mode can be found in
/templates/system/maintainance.tpl.php

Any template can have a custom maintainance template placed in
/templates/yourtemplate/system/maintainance.tpl.php

At this time the maintainance mode is limited to use the default template
set in website options in the backend.

As a next step there will be an admintool that allows to set the mode from
the backend.