Secureform Singletab is gone, i guess i wont be crying.
Singletab is finally removed from Core the file is removed and the only
secureform loaded now is the multi tab variant. I guess noone will be
missing this.

I still need to fix the secureform switcher as the singletab setting
is no longer needed but the fingerprit options are still usefull.