Added hooks for OPF dashboard.
OpF dashboard is a powerfull outputfilter controlling system made by thorn.
Cause of personal trouble it never made it into the core. Not at least its
system hooks are intergrated into the core for as long as we don not have
any other "real" output filter system.

Thanks to mrbaseman :
http://forum.wbce.org/viewtopic.php?id=176