Loginform Autocomplete and Upgrade Script
Small review of upgrade script especially issue #48 and a few other
minor changes. Now using Settings::Set() instead of the function build
into the script as it allows to not overwrite existing Settings.

Allowing loginforms to have autocomplete was regarded a security issue,
So please templatebuilders  pull before  editing the templates as i already
added this once a time ago ;-)