Undefined index: _media_test in /.../admin/media/upload.php on line 103
http://forum.wbce.org/viewtopic.php?pid=2593#p2593

Problem is in mediacenter, because the mediasettings only exist, when at least one time the mediaoptions are stored. So storing the options will end up this notice until an new folder is created.
One idea for solving this would be to check if mediasettings for a folder do exist:
upload.php line 102

[== PHP ==]
if(file_exists($relative.$filename) && isset($pathsettings[$resizepath])) {