New function for admin tools no_page.
I had the problem that i was working on a admin tool that was offering a
file for download. the problem was that calling new admin() produced some
headers and output cause of silly PHPlib.

Now we have an additional post variable "no_page" to suppress all admin
output.  Simply add:

<input type="hidden" name="no_page" value="no_page" />

To the template and you are done.

This was needed to get modules like backup to the standard apeform format