Allow external links and wblinks in page settings now
You are now allowed enter external links or wblinks in the page settings.

If you just enter :
http://www.google.de
or
[wblink2]
in the "filename" field in page settings.

The link is shown in show_menu functions.

This is made because i wanted menulinks that are actually real links,
therefore we needed this functionality in the core first.

But now i really think of hiding this filename field, as it can be
very confusing to plain users.