Admintools loading info.php  and allow for no_page via info.php.
Admintools now load the info.php for additional avaialble information.
This allows for storing additional options in info.php.

On special option is supported now , you can suppress the full WB output in the Backend via
$module_nopage = true;
in info.php

This was needed for Webadmin tool and maybe is needed for other tools too.
That way webadmin now can run under userrights restrictions of WB as a protected include.
For those who like to edit on a full scale texteditor there a several browser plugins
that allow to use you local editor for working on a textarea.