Admintools permissions
qsocial was so kind to send me a nice patch that allows to set
permissions on admin tools. Now only those admin tools that you got
the permission to use appear in admin tool list.
If you try to acces tool you are not allowed via direct url , you are
redirected to admin tool index.