Added class Tool, the basis for moving the BE to consits only of tool modules.
In the first Step class tool does take care of rendering admin tools.
In the second it will render Settings, and in the last Step it will take
care of rendering all BE Pages.

A nice example for this will be thw new Preferences page that is no longer
a real page. In the Upcomming upload it will be a simple call for a
tool module whith class tool.