Added a new field to pages Table "dlink" for a fully functional menulink
After several tests it turned out that a fully functional menulink needs
an extra field. This field opens the option to have real url and wblinks
in the menulink so redirect is no longer necessary , but still possible.

I just add the field in this release as anything else depends on the
modules  Menulink and SM2/3. They will adapt this after this Release.