Finally  Html blocks are added to class Insert
You now can use AddHtml() to insert  extra Html to your templates .
Functionality is the same as AddJs() but you dont have different variants
of html.

This is usefull for dynamic footers , or Warning bars on top of the page
and many other Applications.

MAybe its nice to remember that this functions also in BE templates now.