framework/functions.php  page_link() makes trouble in FE
As reported in  #134

Topics:
when calling a topics detail page in FE, instead of the page just an error occurs:
Fatal error: Call to a member function page_link() on a non-object in /.../framework/functions.php on line 596

This function is needed in FE and BE so i first decided to add an IF THEN , but then remembered that in BE i do a
$wb = $admin;
So a $wb is always available when a $admin is.

So i only changed this back ti $wb.