Replaced new BE functionality whith old (modularity removed)
Calling admintools from whithin a class caused several tools to fail.
Now we first fix the modules and then add the new feature in a 2.0 Version.
2.0 as it breaks something...

For 1.2.x we stick whith the old variant to not create any bad surprises
for users and have updated modules available when changes are applied.

Still there are a lot of new variables and functionality available in
admintools as the script calling the Tools has been enhanced.