Theme overrides for backend.css and backend.js
Now it should be possible to override backend.css and backend.js
via the BE Theme .

Example  if the code moduel had a backend.css and backend.js
you could do a Override in
/templates/YOURTHEME/modules/code/
Just add one or both files there.
You can even override if codde has non of this files.