Added automatic insertion of placeholders for class insert
The new placeholders are now added to templates that do not have em.
This is done by a new output filter  contained inn Class insert.

As its an output filter you can turn it on and off in the opf settings.

If you encounter any trouble whith this please inform me.