Droplets in Backend
From Now on its possible to use Droplets in then BE pages .
Instead of [[yourdroplet]] you have to use [[[yourdroplet]]]
As filtering for [[yourdroplet]] in the backend would replace all
droplets in WYSIWYG fields for example.

In addition the extra <p> tags created by many Editors are filtered out.
<p>[[yourdroplet]]</p> will become

[[yourdroplet]] again.

Not sure if thats a goob idea but we will see.
This is only done in FE display.

Right now all droplets can be used as BE Droplets, but i guess not many
reallly make sense.

I really need to free Droplets from this silly eval , as this slows down
the execution speed drastically.