Rework the Login process first Step
Added Classes WbUser and WbAuth

WbUser is the basic class for doing a complete user management.
In the future  WBUser should be responsible for adding , editing  and storing users.
Theres alreads a lot of functionality , but its not finished yet.

WbAuth is just a static class to do the basic workload of Authentication from
simply everywhere in WB.
Simply call WbAuth::AutoAuthenticate();

And you are done.