Uninstall of core modules mnot allowed
We have so many core modules that are a really bad idea to uninstall as
it could break the whole site or prevent backend settings , or possibly stop you from logging ins.
So we decided to add an extra Parameter to the List in info.php.

$module_level="<user|core>"

Core modules cannot be deinstalled. You need to change the info.php via AFE or
FTP before you can deinstall such modules.