Double Javascript sysvars in FE
The new automatic Placeholders and the old Frontend functions for Javascript collide and
create a double output for the system vars .

To avoid this , the old functions now overwrite the entriy for the new ones
So sysvars only displayed once.

As a result , we found that its necessar to add some new methods to class
I and class Insert , to delete entries.