Added a new method for modules to add stuff to the main template(JS, CSS, HTML, META and more)
Basicallly this is a extension to CSS to head , so modules, droplets, snipets and whatever can add
small blocks to several places in the template.

In the First step a filter adds several placeholder to the template.
Some are for moving content to this location some for easy replacement of
some elements. If those placeholders are already present in the template
they are not set once more. In general its far better to set those
placeholders manually in your template. In addition you can define your
own placeholders if you need to.

```
<!--(PH) TITLE+ -->< title > wieder Anderer titel < / title><!--(PH) TITLE- -->
<!--(PH) META DESC+ --><meta name="description" content="bl bla "/><!--(PH) META DESC- -->
<!--(PH) META KEY+ --><meta name="keywords" content=""/><!--(PH) META KEY- -->

<!--(PH) META HEAD+ -->
<!--(PH) META HEAD- -->

<!--(PH) CSS HEAD TOP+ -->
<!--(PH) CSS HEAD TOP- -->

<!--(PH) JS HEAD TOP+ -->
<!--(PH) JS HEAD TOP- -->

<!--(PH) CSS HEAD BTM+ -->
<!--(PH) CSS HEAD BTM- -->

<!--(PH) JS HEAD BTM+ -->
<!--(PH) JS HEAD BTM- -->

<!--(PH) JS BODY TOP+ -->
<!--(PH) JS BODY TOP- -->

<!--(PH) HTML BODY TOP+ -->
<!--(PH) HTML BODY TOP- -->

<!--(PH) HTML BODY BTM+ -->
<!--(PH) HTML BODY BTM- -->

<!--(PH) JS BODY BTM+ -->
<!--(PH) JS BODY BTM- -->

```
Have a look At the wbce page source in your browser.
Optionaly take a look into the default wbce templates
as they will be updated in one of the next commits.

Now you can move your code into this places by simply echoing

```
<!--(MOVE) CSS HEAD BTM- -->
<style> .heinor1 {} </style>
<!--(END)-->

<!--(MOVE) JS HEAD BTM+ -->
<!-- Just a comment -->
<script>alert('Hello!');</script>
<!--(END)-->

<!--(MOVE) CSS HEAD BTM+ -->
<link href=\"https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css\" rel=\"stylesheet\" >
<!--(END)-->

```

OR replace the replaceablee tags whith

```
<!--(REPLACE) TITLE -->
< title > wieder Anderer titel < / title>
<!--(END)-->

<!--(REPLACE) META DESC -->
<meta name=\"description\" content=\"bl bla \"/>
<!--(END)-->

```
This should come very handy for modules and anything else that wants to bring its own JS, CSS ....