use the Insert class for the keywords and description in simplepagehead
With the Insert class, the page description and keywords are removed
from the page content when the auto placeholders are inserted.
The insert class assumes that it receives this information in its queue
in order to insert it. Therefore, simplepagehead has to use the insert
class and hand over keywords and description instead of inserting this
information in the earlier stage.