implement better brute force protection
we allow 3 failures from a single client ip now. After that we block
for an increasing time span with each wrong attempt. We store that in the
database because an attacker could easily spawn many sessions, so the session
is not a good place to keep track of this. After the blocking time is over
we allow another login attempt, but if it fails, we further increase the
blocking time. After twice the blocking time we reset the attempts.
Also, banned ips are cleared from the database after one week, hoping
that brute force attacks are stopped by authorities in this time frame.