ensure that the database tables are created during update process
under some conditions it may have happened that not all tables were created correctly, so we explicitly call the statements directly from the update script.
Note that we still keep the mechanism in dbsession: it also alters tables if the encoding is not the preferred one.
We also keep the checks on the required tables in the update script in case anything went wrong with creating the tables.
The 'drop table' statements are moved into a separate file, just as a precautious measure not to lose data in case someone changes the call to the sql import in the future and makes a silly mistake. Not to have the drops in the same file is really an easy change and it can prevent us and our users from so much trouble.