menu_link: add redirection type "200"
In fact, do not redirect the browser to this page but change the page_id
under the hood and start processing that page. This is similar to the
section picker, but not only for a single section but for the whole
page.

This is useful for large websites where one page may fit into several
categories, and you want to show it at several places in the menu.
The other redirect types let the browser jump into a different path of
the page_tree. This one just adds a new page with the same content
elsewhere.