Revert "quickfix for opf bug"
This reverts commit 878f42d620208c745a9bb22f8d735ca15c17eaee.

The "bug" (that absolute paths were used) was fixed in opf dashboard
version 1.5.7. Running the update of any version later than this
eliminates all occurrences of the absolute paths in the database.
The prerequisite: those paths are still correct. When migrating
an older installation to a different server may require to edit
the paths and urls manually - or you upgrade first and migrate the
installation afterwards to a new hoster.

If there should still be cases in which the mechanism doesn't work,
the workaround is to uninstall and re-install opf dashboard,
which requires to comment out this line in its info.php like this:
// $module_level       = 'core';